requirements.txt
clc99
Pillow
keyboard
pyinstaller